+++ google2/CHANGELOG.md
@@ -15,8 +15,9 @@ Changed:
 
 Added:
 - Add `detachDisk`
+- Add `streamUploadBlob`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-c0e6762"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-TRAVIS-REPLACE-ME"`
 
 ## 0.10
 Changed: